S3T-420-Fix : 여행 On 작성 화면 한 페이지로 수정
